Parlay: Factor out Builder into separate file
- Removing lines was causing Builder to reset with every re-render
